reduce float32 to float16

+++ model.py
@@ -35,7 +35,7 @@ class AgenderNetVGG16(Model):
 
 	@staticmethod
 	def prepImg(data):
-		data = data.astype('float32')
+		data = data.astype('float16')
 		data = data[..., ::-1]
 		mean = [103.939, 116.779, 123.68]
 		data[..., 0] -= mean[0]
@@ -67,7 +67,7 @@ class AgenderNetInceptionV3(Model):
 
 	@staticmethod
 	def prepImg(data):
-		data = data.astype('float32')
+		data = data.astype('float16')
 		data /= 127.5
 		data -= 1.
 		return data
@@ -96,7 +96,7 @@ class AgenderNetXception(Model):
 
 	@staticmethod
 	def prepImg(data):
-		data = data.astype('float32')
+		data = data.astype('float16')
 		data /= 127.5
 		data -= 1.
 		return data
